--- a/Makefile
+++ b/Makefile
@@ -1,3 +1,11 @@
+generate:
+	@for name in internal/re2go/*.re; do \
+		RE_IN=$$name; \
+		RE_OUT=$$(echo $$name | sed 's/\.re/.go/'); \
+		re2go -W -F --input-encoding utf8 --utf8 --no-generation-date -i $$RE_IN -o $$RE_OUT; \
+		gofmt -w $$RE_OUT; \
+	done
+
 test-all:
 	@echo "Test normal regex"
 	@echo