--- a/constant.go
+++ b/constant.go
@@ -23,8 +23,11 @@ import (
 	"time"
 
 	dps "github.com/markusmobius/go-dateparser"
+	"github.com/markusmobius/go-htmldate/internal/re2go"
 	"github.com/markusmobius/go-htmldate/internal/regexp"
 )
+
+type fnRe2GoFinder func(string) [][]int
 
 var (
 	timeZero       = time.Time{}
@@ -127,32 +130,33 @@ var (
 	// `([0-9]{1,2})[./]([0-9]{1,2})[./]([0-9]{2,4}) *?(?:'de|'da|'te|'ta|’de|’da|’te|’ta|tarihinde) *(?:güncellendi|yayı(?:m|n)landı)`) // TR
 
 	// Extensive search patterns
-	rxYearPattern       = compileRegexF(`^\D?(%s)`, rxYear)
-	rxCopyrightPattern  = compileRegexF(`(?:©|\&copy;|Copyright|\(c\))\D*(?:%[1]s-)?(%[1]s)\D`, rxYear)
-	rxThreePattern      = regexp.MustCompile(`/([0-9]{4}/[0-9]{2}/[0-9]{2})[01/]`)
-	rxThreeCatch        = regexp.MustCompile(`([0-9]{4})/([0-9]{2})/([0-9]{2})`)
-	rxThreeLoosePattern = regexp.MustCompile(`\D([0-9]{4}[/.-][0-9]{2}[/.-][0-9]{2})\D`)
-	rxThreeLooseCatch   = regexp.MustCompile(`([0-9]{4})[/.-]([0-9]{2})[/.-]([0-9]{2})`)
+	rxYearPattern = compileRegexF(`^\D?(%s)`, rxYear)
+	// rxCopyrightPattern  = compileRegexF(`(?:©|\&copy;|Copyright|\(c\))\D*(?:%[1]s-)?(%[1]s)\D`, rxYear)
+	// rxThreePattern      = regexp.MustCompile(`/([0-9]{4}/[0-9]{2}/[0-9]{2})[01/]`)
+	rxThreeCatch = regexp.MustCompile(`([0-9]{4})/([0-9]{2})/([0-9]{2})`)
+	// rxThreeLoosePattern = regexp.MustCompile(`\D([0-9]{4}[/.-][0-9]{2}[/.-][0-9]{2})\D`)
+	rxThreeLooseCatch = regexp.MustCompile(`([0-9]{4})[/.-]([0-9]{2})[/.-]([0-9]{2})`)
 	// rxSelectYmdPattern   = regexp.MustCompile(`\D([0-3]?[0-9][/.-][01]?[0-9][/.-][0-9]{4})\D`)
-	rxSelectYmdYear      = compileRegexF(`(%s)\D?$`, rxYear)
-	rxYmdYear            = compileRegexF(`^(%s)`, rxYear)
-	rxDateStringsPattern = regexp.MustCompile(`(\D19[0-9]{2}[01][0-9][0-3][0-9]\D|\D20[0-9]{2}[01][0-9][0-3][0-9]\D)`)
-	rxDateStringsCatch   = compileRegexF(`(%s)([01][0-9])([0-3][0-9])`, rxYear)
+	rxSelectYmdYear = compileRegexF(`(%s)\D?$`, rxYear)
+	rxYmdYear       = compileRegexF(`^(%s)`, rxYear)
+	// rxDateStringsPattern = regexp.MustCompile(`(\D19[0-9]{2}[01][0-9][0-3][0-9]\D|\D20[0-9]{2}[01][0-9][0-3][0-9]\D)`)
+	rxDateStringsCatch = compileRegexF(`(%s)([01][0-9])([0-3][0-9])`, rxYear)
 	// rxSlashesPattern     = regexp.MustCompile(`\D([0-3]?[0-9]/[01]?[0-9]/[0129][0-9]|[0-3][0-9]\.[01][0-9]\.[0129][0-9])\D`)
-	rxSlashesYear   = regexp.MustCompile(`([0-9]{2})$`)
-	rxYyyyMmPattern = regexp.MustCompile(`\D([12][0-9]{3}[/.-](?:1[0-2]|0[1-9]))\D`)
-	rxYyyyMmCatch   = compileRegexF(`(%s)[/.-](1[0-2]|0[1-9])`, rxYear)
+	rxSlashesYear = regexp.MustCompile(`([0-9]{2})$`)
+	// rxYyyyMmPattern = regexp.MustCompile(`\D([12][0-9]{3}[/.-](?:1[0-2]|0[1-9]))\D`)
+	rxYyyyMmCatch = compileRegexF(`(%s)[/.-](1[0-2]|0[1-9])`, rxYear)
 	// rxMmYyyyPattern   = regexp.MustCompile(`\D([01]?[0-9][/.-][12][0-9]{3})\D`)
-	rxMmYyyyYear      = compileRegexF(`(%s)\D?$`, rxYear)
-	rxSimplePattern   = compileRegexF(`\D(%s)\D`, rxYear)
+	rxMmYyyyYear = compileRegexF(`(%s)\D?$`, rxYear)
+	// rxSimplePattern   = compileRegexF(`\D(%s)\D`, rxYear)
 	rxSimpleW3Cleaner = compileRegexF(`w3.org\D(%s)\D`, rxYear)
 
 	rxThreeComponents = []struct {
-		Pattern *regexp.Regexp
+		Name    string
+		Pattern fnRe2GoFinder
 		Catcher *regexp.Regexp
 	}{
-		{rxThreePattern, rxThreeCatch},
-		{rxThreeLoosePattern, rxThreeLooseCatch},
+		{"ThreePattern", re2go.ThreePattern, rxThreeCatch},
+		{"ThreeLoosePattern", re2go.ThreeLoosePattern, rxThreeLooseCatch},
 	}
 
 	// Time patterns