--- a/constant.go
+++ b/constant.go
@@ -19,12 +19,11 @@ package htmldate
 
 import (
 	"fmt"
-	stdregex "regexp"
+	"regexp"
 	"time"
 
 	dps "github.com/markusmobius/go-dateparser"
 	"github.com/markusmobius/go-htmldate/internal/re2go"
-	"github.com/markusmobius/go-htmldate/internal/regexp"
 )
 
 type fnRe2GoFinder func(string) [][]int
@@ -117,8 +116,8 @@ var (
 
 	// Time patterns
 	rxCommonTime = regexp.MustCompile(`(?i)(?:\D|^)(\d{1,2})(?::|\s*h\s*)(\d{1,2})(?::(\d{1,2})(?:\.\d+)?)?(?:\s*((?:a|p)\.?m\.?))?`)
-	rxTzCode     = stdregex.MustCompile(`(?i)(?:\s|^)([-+])(\d{2})(?::?(\d{2}))?`)
-	rxIsoTime    = stdregex.MustCompile(`(?i)(\d{2}):(\d{2})(?::(\d{2})(?:\.\d+)?)?(Z|[+-]\d{2}(?::?\d{2})?)`)
+	rxTzCode     = regexp.MustCompile(`(?i)(?:\s|^)([-+])(\d{2})(?::?(\d{2}))?`)
+	rxIsoTime    = regexp.MustCompile(`(?i)(\d{2}):(\d{2})(?::(\d{2})(?:\.\d+)?)?(Z|[+-]\d{2}(?::?\d{2})?)`)
 
 	rxLastJsonBracket = regexp.MustCompile(`(?i)\s*\}$`)
 )