--- a/internal/re2go/idiosyncrasies-search.go
+++ b/internal/re2go/idiosyncrasies-search.go
@@ -14,7 +14,7 @@ func IdiosyncracyPatternSubmatch(input string) ([]string, int) {
 	_ = marker
 
 	// Variable for capturing parentheses (twice the number of groups).
-	var YYMAXNMATCH int = 4
+	const YYMAXNMATCH = 4
 
 	yypmatch := make([]int, YYMAXNMATCH*2)
 	var yynmatch int