… README.md
… core_test.go
… go.mod
… go.sum
… internal/re2go/base.go
… internal/re2go/base.re
… internal/re2go/idiosyncrasies-search.re
… internal/re2go/pattern.go
… internal/re2go/pattern.re
… internal/re2go/plausible-year.go
… internal/re2go/plausible-year.re
… internal/re2go/re2go_test.go
… internal/re2go/regex-parse.go
… internal/re2go/regex-parse.re
… internal/re2go/regex-pattern-search.go
… internal/re2go/regex-pattern-search.re
… internal/re2go/search-pattern-test.go
… internal/re2go/search-pattern-test.re
… internal/re2go/search-pattern.go
… internal/re2go/search-pattern.re
… internal/re2go/utils.go
… internal/regexp/regexp-re2.go
… internal/regexp/regexp.go
… utils.go
… validators.go