## What is it?

I was working on go-trafilatura and noticed that htmldate actually is the main bottleneck for go-trafilatura. After running `scripts/comparison` with profiler, go-htmldate able to process 1000 web pages in 6.21 seconds, which is pretty fast (only 6ms per page). However it's in fact is the main bottleneck for go-trafilatura so there are needs to improve it. Here is the profile result:

![The old go-htmldate](https://github.com/user-attachments/assets/54212124-968c-42ef-acaf-aaa995711d7d)

From that graph above, in a glance we can see that regular expressions took most of the time, so I filter it to get the exact number:

![How much regex we use in old version](https://github.com/user-attachments/assets/754065e8-3fdf-432d-91be-430ebf30f4df)

So, from the entire 6.21 seconds, regular expressions alone use 5.04 seconds or around 81% of the entire process.

We already know that Go's regex is [slow](https://github.com/golang/go/issues/26623#issuecomment-1899965448), so we already use [`github.com/wasilibs/go-re2`](https://github.com/wasilibs/go-re2) package that uses cgo. However cgo make it hard to cross compile Go' application.

After looking around, I found about [re2go](https://re2c.org/manual/manual_go.html#regular-expressions). It's a compiler that compile regular expressions ahead of time into Go' code. It's pretty interesting:

- It uses a new-ish lookahead TDFA algorithm, which supposed to be faster than Go' regex that use NFA and one-pass algorithm.
- It has a long history and is still actively maintained.
- It's already used in production by various open source projects like PHP and Ninja.
- After the initial learning curve, it's actually pretty easy to use.
- From [my regex benchmark](https://github.com/RadhiFadlillah/go-regex-benchmark-web/tree/concurrent), for some case it's actually faster than `wasilibs/go-re2` that uses cgo binding.

## Which regex to replace?

Obviously we can't replace every regex with re2go, so I looked at the profile to see which regex to replace. I found several regex that take most of the time:

1. First is `rxIdiosyncrasy` that used in `extractIdiosyncrasy` function. That one regex alone take around 2.65 seconds.
2. Next is `rxLongTextPattern` that used in `regexParse` function. That one regex take around 0.7 seconds.
3. Next is `rxTimestampPattern` that used in `regexPatternSearch` function. That one regex take around 0.65 seconds.
4. Next are `rxSelectYmdPattern`, `rxSlashesPattern`, `rxMmYyyyPattern`, `rxCopyrightPattern`, `rxThreePattern`, `rxThreeLoosePattern`, `rxDateStringsPattern`, `rxYyyyMmPattern` and `rxSimplePattern`. All of them are used by `plausibleYearFilter` and take around 0.4 seconds.

All together take around 4.4 seconds, which is around 87% of the entire regex processes.

## The result

Here is the new profile after converting those regex with re2go:

![The result](https://github.com/user-attachments/assets/00d24932-1fb6-4bae-b09c-90d91858cd60)

Now go-htmldate able to process 1000 web pages in 1.49 seconds, or 4x faster than the old one. The best thing is it's still pure Go code without cgo binding, so it's still easy to cross compile.

There are still leftover regex, but it's all from the external `go-dateparser` package:

![The leftover regex](https://github.com/user-attachments/assets/974a1071-bffb-44c2-902f-b6331fecaec2)

## Additional notes

Besides converting those regex I also added tests for the code generated by re2go. I also updated all dependencies to the latest one.

If you want to see the profiles yourself, you can download it here: [htmldate-profile.zip](https://github.com/user-attachments/files/17044444/htmldate-profile.zip)